safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-edgeMode-prop-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: SVGFEConvolveMatrixElement
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'edgeMode' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
